Make sure to always have the Frame python wrapper when imported
- Make tests actuallyt test that
