+++ python/podio/__init__.py
@@ -1,6 +1,4 @@
 """Python module for the podio EDM toolkit and its utilities"""
-import sys
-
 from .__version__ import __version__
 
 import ROOT  # pylint: disable=wrong-import-order
@@ -14,7 +12,7 @@ _DYNAMIC_LIBS_LOADED = False
 # silence any ouptput. If we can find it, we can also safely load it
 if ROOT.gSystem.DynamicPathName("libpodioDict.so", True):
   ROOT.gSystem.Load("libpodioDict.so")
-  from ROOT import podio
+  from ROOT import podio  # noqa: F401
 
   _DYNAMIC_LIBS_LOADED = True
 
@@ -31,18 +29,6 @@ if _DYNAMIC_LIBS_LOADED:
     pass
 
   from . import EventStore
-
-  try:
-    # For some reason the test_utils only work at (test) runtime if they are
-    # imported with the rest of podio. Otherwise they produce a weird c++ error.
-    # This happens even if we import the *exact* same file.
-    from . import test_utils  # noqa: F401
-  except ImportError:
-    pass
-
-  # Make sure that this module is actually usable as podio even though most of
-  # it is dynamically populated by cppyy
-  sys.modules["podio"] = podio
 
   __all__ = [
       "__version__",
